Skill ut SecurityConfig.kt til egen fil Lag egen SecurityConfig for dev-profil som åpner alle endepunkt
