+++ src/main/kotlin/no/nav/familie/ks/barnehagelister/rest/DefaultBarnehagelisteController.kt
@@ -10,12 +10,14 @@ import no.nav.familie.ks.barnehagelister.rest.dto.toResponseEntity
 import no.nav.familie.ks.barnehagelister.service.BarnehagelisteService
 import org.springframework.context.annotation.Profile
 import org.springframework.http.ResponseEntity
+import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.validation.BindingResult
 import org.springframework.validation.FieldError
 import org.springframework.web.bind.annotation.RestController
 import java.util.UUID
 
 @Profile("!dev")
+@PreAuthorize("hasAuthority('SCOPE_nav:familie/v1/kontantstotte/barnehagelister')")
 @RestController
 class DefaultBarnehagelisteController(
     private val barnehagelisteService: BarnehagelisteService,
